Move retry logic to its own file, retryLogic.ts
