+++ ts/packages/agents/spelunker/src/searchCode.ts
@@ -32,6 +32,7 @@ import {
 } from "./embeddings.js";
 import { OracleSpecs } from "./oracleSchema.js";
 import { chunkifyPythonFiles } from "./pythonChunker.js";
+import { retryOn429 } from "./retryLogic.js";
 import { ChunkDescription, SelectorSpecs } from "./selectorSchema.js";
 import { SpelunkerContext } from "./spelunkerActionHandler.js";
 import { SummarizerSpecs } from "./summarizerSchema.js";
@@ -368,7 +369,7 @@ async function selectRelevantChunks(
     ${prepareChunks(chunks)}
     `;
     // console_log(prompt);
-    const result = await retryTranslateOn429(() => selector.translate(prompt));
+    const result = await retryOn429(() => selector.translate(prompt));
     if (!result) {
         console_log(`  [Failed to select chunks for ${chunks.length} chunks]`);
         return [];
@@ -690,9 +691,7 @@ async function summarizeChunkSlice(
     ${prepareChunks(chunks)}
     `;
     // console_log(prompt);
-    const result = await retryTranslateOn429(() =>
-        summarizer.translate(prompt),
-    );
+    const result = await retryOn429(() => summarizer.translate(prompt));
     if (!result) {
         console_log(
             `  [Failed to summarize chunks for ${chunks.length} chunks]`,
@@ -759,58 +758,6 @@ async function summarizeChunkSlice(
         }
     }
     if (errors) console_log(`  [${errors} errors]`);
-}
-
-export async function retryTranslateOn429<T>(
-    translate: () => Promise<Result<T>>,
-    retries: number = 3,
-    defaultDelay: number = 5000,
-): Promise<T | undefined> {
-    let wrappedResult: Result<T>;
-    do {
-        retries--;
-        wrappedResult = await translate();
-        // console_log(wrappedResult);
-        if (!wrappedResult.success) {
-            if (
-                retries > 0 &&
-                wrappedResult.message.includes("fetch error: 429:")
-            ) {
-                let delay = defaultDelay;
-                const embeddingTime = wrappedResult.message.match(
-                    /Try again in (\d+) seconds/,
-                );
-                const azureTime = wrappedResult.message.match(
-                    /after (\d+) milliseconds/,
-                );
-                const openaiTime = wrappedResult.message.match(
-                    /Please try again in (\d+\.\d*|\.\d+|\d+m)s./,
-                );
-                if (embeddingTime || azureTime || openaiTime) {
-                    if (embeddingTime) {
-                        delay = parseInt(embeddingTime[1]) * 1000;
-                    } else if (azureTime) {
-                        delay = parseInt(azureTime[1]);
-                    } else if (openaiTime) {
-                        delay = parseFloat(openaiTime[1]);
-                        if (!openaiTime[1].endsWith("m")) {
-                            delay *= 1000;
-                        }
-                    }
-                } else {
-                    console_log(
-                        `      [Couldn't find msec in '${wrappedResult.message}'`,
-                    );
-                }
-                console_log(`    [Retry on 429 error: sleep ${delay} ms]`);
-                await new Promise((resolve) => setTimeout(resolve, delay));
-                continue;
-            }
-            console_log(`    [Giving up: ${wrappedResult.message}]`);
-            return undefined;
-        }
-    } while (!wrappedResult.success);
-    return wrappedResult.data;
 }
 
 function keepBestChunks(
